move filter into congestion control
